requirements.txt
pyTelegramBotAPI==4.14.0
solana==0.30.2
aiohttp==3.9.1
asyncio-throttle==1.0.2
python-dotenv==1.0.0
